@@ -11,6 +11,7 @@
 
 namespace App\Services\Report;
 
+use App\Models\User;
 use App\Utils\Ninja;
 use App\Utils\Number;
 use App\Models\Client;
@@ -21,6 +22,7 @@ use App\Libraries\MultiDB;
 use App\Export\CSV\BaseExport;
 use App\Utils\Traits\MakesDates;
 use Illuminate\Support\Facades\App;
+use App\Services\Template\TemplateService;
 
 class ClientBalanceReport extends BaseExport
 {
@@ -32,6 +34,10 @@ class ClientBalanceReport extends BaseExport
     public Writer $csv;
 
     public string $date_key = 'created_at';
+
+    private string $template = '/views/templates/reports/client_balance_report.html';
+
+    private array $clients = [];
 
     public array $report_keys = [
         'client_name',
@@ -107,6 +113,32 @@ class ClientBalanceReport extends BaseExport
         return $headers;
 
     }
+
+    public function getPdf()
+    {
+        $user = isset($this->input['user_id']) ? User::withTrashed()->find($this->input['user_id']) : $this->company->owner();
+
+        $user_name = $user ? $user->present()->name() : '';
+
+        $data = [
+            'clients' => $this->clients,
+            'company_logo' => $this->company->present()->logo(),
+            'company_name' => $this->company->present()->name(),
+            'created_on' => $this->translateDate(now()->format('Y-m-d'), $this->company->date_format(), $this->company->locale()),
+            'created_by' => $user_name,
+        ];
+
+        $ts = new TemplateService();
+
+        $ts_instance = $ts->setCompany($this->company)
+                    ->setData($data)
+                    ->setRawTemplate(file_get_contents(resource_path($this->template)))
+                    ->parseNinjaBlocks()
+                    ->save();
+
+        return $ts_instance->getPdf();
+    }
+
     private function buildRow(Client $client): array
     {
         $query = Invoice::query()->where('client_id', $client->id)
@@ -114,7 +146,7 @@ class ClientBalanceReport extends BaseExport
 
         $query = $this->addDateRange($query, 'invoices');
 
-        return [
+        $item = [
             $client->present()->name(),
             $client->number,
             $client->id_number,
@@ -123,5 +155,9 @@ class ClientBalanceReport extends BaseExport
             Number::formatMoney($client->credit_balance, $this->company),
             Number::formatMoney($client->payment_balance, $this->company),
         ];
+
+        $this->clients[] = $item;
+
+        return $item;
     }
 }
