@@ -1099,7 +1099,7 @@ $lang = array(
     'invoice_embed_documents' => 'Embed Documents',
     'invoice_embed_documents_help' => 'Include attached images in the invoice.',
     'document_email_attachment' => 'Attach Documents',
-    'ubl_email_attachment' => 'Attach UBL/E-Invoice',
+    'ubl_email_attachment' => 'Attach UBL',
     'download_documents' => 'Download Documents (:size)',
     'documents_from_expenses' => 'From Expenses:',
     'dropzone_default_message' => 'Drop files or click to upload',
